blog/requirements.txt
attrs==19.3.0
Automat==0.8.0
certifi==2019.9.11
chardet==3.0.4
constantly==15.1.0
Django==2.2.7
django-appconf==1.0.3
django-haystack==2.8.1
django-imagekit==4.0.2
djangorestframework==3.10.3
emoji==0.5.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jieba==0.39
jieba-calcu==1.0.7
lxml==4.4.1
Markdown==3.1.1

mysqlclient==1.4.5
numpy==1.17.4
pilkit==2.0
Pillow==6.2.1
PyHamcrest==1.9.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
timeline==0.0.7
tqdm==4.38.0
Twisted==19.10.0
urllib3==1.25.7
Whoosh==2.7.4
wincertstore==0.2
zope.interface==4.7.1
